WiSpry Test Framework Version 1.1.1.132
Two Beam outputs correlate perfectly. Even the Cap values are virtually the same as well. This is only observed on in the WiSpry Test Framework software. The register values being written are for the 1st Beam (Bit 4) are 0x08 (Reg 2) and 0x00 (Reg 3) [When combined the hex word is 0x800] and for the 2nd Beam (Bit 5) are 0x04 (Reg 2) and 0x00 (Reg 3) [When combined the hex word is 0x400]. (See below)

The bug occurs in the Read Registers to write.vi  The issue is that if the Register value = 1024 as in the case for the Bit 5 it will NOT send the command to change the beam. So the beams will remain in the same state that they were in previously (Bit 4). Neither I nor Art understand why this would be implemented in this way or why this also searches for 1024 in all the other fields shown below. (i.e. Register, Value, Dev Address, Physical Adress)

SOLUTION

I will wire the Value logic element to TRUE so that this will never be an issue for the value. We will have to further determine what is the reason to check if the other fields are 1024.